Explode Groovy files, each class own file
