+++ gradle/plugins-inventory/src/main/groovy/GenerateDebugMappingPlugin.groovy
