+++ gradle/plugins-inventory/src/main/groovy/InventoryDatabaseEntity.groovy
@@ -1,0 +1,12 @@
+
+class InventoryDatabaseEntity {
+	File input
+	File output
+	File iconFolder
+	String conversion
+
+	final String name
+	InventoryDatabaseEntity(String name) {
+		this.name = name
+	}
+}
