+++ gradle/plugins-inventory/src/main/groovy/InventoryDatabasePlugin.groovy
@@ -44,15 +44,3 @@ class InventoryDatabasePlugin implements Plugin<Project> {
 		}
 	}
 }
-
-class InventoryDatabaseEntity {
-	File input
-	File output
-	File iconFolder
-	String conversion
-
-	final String name
-	InventoryDatabaseEntity(String name) {
-		this.name = name
-	}
-}
